Worked on the header of the page.
Added some images to the image folder. Added a container to enclose the entire content of the page, inside the body. Created a header to wrap contents at the top of the page. Added a background image to the header and styled it. Made the header responsive by including some media query for small devices.